users/requirements.txt
fastapi ~= 0.66.0 
uvicorn ~= 0.14
python-dotenv ~= 0.18.0
pydantic ~= 1.8.2 
pytest ~= 6.2.4
pylint ~= 2.12
SQLAlchemy ~= 1.4.21 
alembic ~= 1.6.5 
passlib ~= 1.7
python-jose[cryptography] ~= 3.3
python-multipart ~= 0.0.5 
bcrypt ~= 3.2.0 
